La til nye modeller for Kepler 1

+++ HenonHeiles.py
@@ -106,7 +106,7 @@ class HenonHeilesSystem:
     
 
 def generate_data(ntrajectories, t_sample,system= HenonHeilesSystem(),integrator = "RK4",true_derivatives = False,H0=None,u0s=None):
-    data_type = torch.float64
+    data_type = torch.float32
     #Parameters
     nstates = system.nstates
     traj_length = t_sample.shape[0]
