+++ numerical_integration.py
@@ -3,7 +3,7 @@ import torch
 from scipy.optimize import newton
 
 def newton_torch(func, guess, threshold=1e-7, max_iters=1000, damping=1.0):
-    guess = torch.tensor(guess, dtype=torch.float64, requires_grad=True)
+    guess = torch.tensor(guess, dtype=torch.float32, requires_grad=True)
     for i in range(max_iters):
         value = func(guess) 
         if torch.linalg.norm(value) < threshold: #Converged
